+++ src/Helper/XMLParser.php
@@ -260,11 +260,10 @@ class XMLParser
             for ($offset = 0; $offset < $len; $offset += $this->maxChunkLength) {
                 $chunk = substr($data, $offset, $this->maxChunkLength);
                 // error handling: xml not well formed
-                if (!xml_parse($parser, $chunk, $offset + $this->maxChunkLength >= $len)) {
+                if (!@xml_parse($parser, $chunk, $offset + $this->maxChunkLength >= $len)) {
                     $errCode = xml_get_error_code($parser);
                     $errStr = sprintf('XML error %s: %s at line %d, column %d', $errCode, xml_error_string($errCode),
                         xml_get_current_line_number($parser), xml_get_current_column_number($parser));
-
                     $this->_xh['isf'] = 3;
                     $this->_xh['isf_reason'] = $errStr;
                 }
@@ -780,6 +779,9 @@ class XMLParser
                 } elseif ($this->_xh['isf'] == 0 && count($this->_xh['params']) !== 1) {
                     $this->_xh['isf'] = 2;
                     $this->_xh['isf_reason'] = "PARAMS element inside METHODRESPONSE should have exactly 1 PARAM";
+                } elseif ($this->_xh['isf'] == 1 && $this->_xh['params'] !== false) {
+                    $this->_xh['isf'] = 2;
+                    $this->_xh['isf_reason'] = "both FAULT and PARAMS elements found inside METHODRESPONSE";
                 }
                 break;